requirements.txt
beautifulsoup4
flask
openpyxl
pandas
psycopg2-binary
python-dotenv
requests
sqlalchemy
sqlalchemy-cockroachdb
